Split functionality into classes + documentation
- Removed PooledVideo inner class
- Removed user interning functionality to UserPool
- Added lots of comments
